Duncan Craft 2000 v1.3.0B
This is a big update so I may forget some of the things I added
•	Game now runs on primarily a tick system with some threading
•	New Minecraft style lighting
•	New torch block
•	Added in random ticks, so sometimes dirt blocks if adjacent to a grass block will become a grass block
•	Early implementation of biomes
•	Fog
•	A small reorganization of the codebase
•	Added in so worlds will sort by last played and not alphabetically and you can now press del and it will delete the currently selected world in the title screen
•	Probably added in a memory leak
•	Fixed some bug
This update has taken a while to get done how I liked it. Apparently adding complex lighting to a voxel game is really hard. It’s still not 100% of what I like but it works and that’s fine by me.